% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fancy_scientific.R
\name{fancy_scientific}
\alias{fancy_scientific}
\title{Fancy Scientific Notation}
\usage{
fancy_scientific(l)
}
\arguments{
\item{l}{Input label}
}
\value{
The number value in scientific notation
}
\description{
Code taken from \url{http://stackoverflow.com/questions/11610377/how-do-i-change-the-formatting-of-numbers-on-an-axis-with-ggplot}
}

