% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cols_by_names.R
\name{get_cols_by_names}
\alias{get_cols_by_names}
\title{Get Columns of Matrix by Column Name}
\usage{
get_cols_by_names(in.mat, in.col.names)
}
\arguments{
\item{in.mat}{The input matrix}

\item{in.col.names}{The column names to subset by}
}
\value{
A subset of the matrix
}
\description{
This function provides a pipe friendly way of retrieving columns of a matrix
by column anmes
}
\examples{
mat <- matrix(1:20, 5, 4, dimnames = list(NULL, LETTERS[1:5]))
get_cols_by_names(mat, c("A", "D", "E"))
}

