% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_legend.R
\name{get_legend}
\alias{get_legend}
\title{Get the Legend for ggplot}
\usage{
get_legend(in.ggplot)
}
\arguments{
\item{in.ggplot}{Input ggplot with legend}
}
\value{
ggplot legend
}
\description{
Retrieve the legend of a ggplot so that it can be used as a plot itself.
This is useful in situations where you are arrange multiple ggplots and want
to share a common legend.
}
\author{
\url{http://www.sthda.com/english/wiki/ggplot2-easy-way-to-mix-multiple-graphs-on-the-same-page-r-software-and-data-visualization}
}

