% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lower_upper_tri.R
\name{get_lower_upper_tri}
\alias{get_lower_upper_tri}
\title{Get Triangle of the Correlation Matrix}
\usage{
get_lower_upper_tri(cor.mat, tri.type = c("none", "lower", "upper"))
}
\arguments{
\item{cor.mat}{Correlation matrix.}

\item{tri.type}{Retrieve values below or above the matrix. If 'none' is 
specified, then the same input matrix is returned.}
}
\value{
matrix with values above or below the diagonal depending on the 
  tri.type.
}
\description{
Gets a triangle (lower or upper) of a correlation matrix.
}

