% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_rows_by_names.R
\name{get_rows_by_names}
\alias{get_rows_by_names}
\title{Get Row of Matrix by Rowname}
\usage{
get_rows_by_names(in.mat, in.row.names)
}
\arguments{
\item{in.mat}{The input matrix}

\item{in.row.names}{The row names to subset by}
}
\value{
A subset of the matrix
}
\description{
This function provides a pipe friendly way of retrieving rows of a matrix
by rownames.
}
\examples{
mat <- matrix(1:20, 5, 4, dimnames = list(LETTERS[1:5], NULL))
get_row_by_name(mat, c("A", "D", "E"))
}

