% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snv_read_count.R
\name{get_snv_read_count}
\alias{get_snv_read_count}
\title{Get Nucleotide Read Counts}
\usage{
get_snv_read_count(bamfile, scan.bam.param, pileup.param)
}
\arguments{
\item{bamfile}{Path to the input bam file.}

\item{pileup.param}{Rsamtools::PileupParam object which specifies specific
pileup parameters for filtering.}

\item{scan.bam.params}{Rsamtools::ScanBamParam object which specifies 
positions to lookup and additional read filtering. See the ScanBamParam 
documentation for specific details.}
}
\value{
data.frame containing the nucleotide and read count information for each
  position specified.
}
\description{
Wrapper function over Rsamtools::pileup to retrieve nucleotide read count 
values over a set of positions. This function serves more as a way to 
document how to use the pileup function.
}
\examples{
library("GenomicRanges")
library("Rsamtools")

bamfile <- system.file("extdata", "ex1.bam", package = "Rsamtools")

# Setup Lookup Positions 
positions.df <- data.frame(chr = c("seq1", "seq1", "seq1"),
                           start = c(50, 891, 1000),
                           stop = c(50, 891, 1000))
positions.gr <- makeGRangesFromDataFrame(positions.df)
scan.bam.param <- ScanBamParam(which = positions.gr)

# Setup Read Filters
pileup.param <- PileupParam(max_depth = 1000, min_mapq = 13, 
                            min_base_quality = 20, 
                            distinguish_strands = FALSE, 
                            min_nucleotide_depth = 0)

get_snv_read_count(bamfile, scan.bam.param)
}

