% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Get the Default ggplot2 Color Palette}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Number of Coolors}
}
\description{
\code{gg_color_hue} returns the default ggplot color palette for a given
number of colors
}
\author{
John Colby \url{http://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette}
}

