% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_presence_absence_mat.R
\name{make_PA_matrix}
\alias{make_PA_matrix}
\title{Create a presence/asbence (PA) matrix}
\usage{
make_PA_matrix(inList)
}
\arguments{
\item{inList}{List of vectors containing elements.}
}
\value{
A matrix containing various vector elements as rows and the 
 names in the list as columns
}
\description{
This function given a list of vectors containing various elements 
(e.g. gene names) will create a matrix with the elements as rows and 
each vector as a column. TRUE/FALSE values wil indicate whether the 
element was found part of the vector
}
\examples{
v1 <- c("EZH2", "B2M", "RCOR1", "CREBBP")
v2 <- c("LCOR", "NCOR1", "MLL2", "CREBBP", "CD58")
v3 <- c("TP53", "UNC80", "CREBBP")
inList <- list("groupA" = v1, "groupB" = v2, "groupC" = v3)
make_PA_matrix(inList)
}

