% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_presence_absence_mat.R
\name{MakePAMatrix}
\alias{MakePAMatrix}
\title{Deprecated Function}
\usage{
MakePAMatrix(inList)
}
\description{
This function is deprecated in favour of renaming it to make_PA_matrix
}

