% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mat.R
\name{format_mat}
\alias{format_mat}
\title{Format Matrix}
\usage{
format_mat(mat, log.transform = FALSE, log.base = 2,
  norm.method = c("none", "quantile", "z", "max"), min.val)
}
\arguments{
\item{mat}{The input matrix.}

\item{log.transform}{Boolean to indicate whether the log transform the
matrix.}

\item{log.base}{The base to use for the log transformation.}

\item{min.val}{Minimum value of the matrix. Any values in the matrix below
this value will be set to this value. This is useful if trying to log
transform a matrix. For instance, if log2 transforming then any values
less than 1 will produce negative values. Setting min.val = 1 will ensure
that any values less than 1 be log transformed to 0.}

\item{correct.inf.flag}{Booelean to set whether.}
}
\value{
A formatted matrix.
}
\description{
This function provides the ability to format your matrix in various ways.
For instance, you can log transform, row normalize, and quantile normalize.
It is best to use the log.transform and norm.method separately (i.e. make to
calls of format_mat)
}

