% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format_mat.R
\name{format_mat}
\alias{format_mat}
\title{Format Matrix}
\usage{
format_mat(mat, log.transform = FALSE, log.base = 2,
  norm.method = c("none", "quantile", "z", "max"))
}
\arguments{
\item{mat}{The input matrix}

\item{log.transform}{Boolean to indicate whether the log transform the
matrix}

\item{log.base}{The base to use for the log transformation}

\item{correct.inf.flag}{Booelean to set whether}
}
\value{
A formatted matrix
}
\description{
This function provides the ability to format your matrix in various ways.
For instance, you can log transform, row normalize, and quantile normalize.
It is best to use the log.transform and norm.method separately (i.e. make to
calls of format_mat)
}

