% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_feature_sample_mat.R
\name{plot_feature_sample_mat}
\alias{plot_feature_sample_mat}
\title{Plot a Feature-Sample Matrix}
\usage{
plot_feature_sample_mat(in.dt, feature.order, sample.id.order, fill.colors,
  type.display.mode = c("multiple", "single"), type.order, tile.col,
  rotate.x.labels)
}
\arguments{
\item{in.dt}{A 3 column (feature, sampleID, type) data.table object}

\item{feature.order}{character vector indicating the order of the features
in the final plot on the y-axis. If not set, then function will set it
automatically}

\item{sample.id.order}{character vector indicating the order of the samples
in the final plot on the x-axis. If not set, then function will set it
automatically}

\item{fill.colors}{character vector indicating the colors of the different
"types". The names should be the types with the value being the color}

\item{type.display.mode}{Specify whether multiple or a single feature type
can appear in the same feature/sample cell}

\item{type.order}{Specify the "priority" of the feature types. This only
has an effect when type.display.mode is set to single}

\item{tile.col}{Border color of each cell. If not yet, no border color is
used}

\item{rotate.x.labels}{Rotate the x-axes labels by a certain degree}
}
\description{
Generates a ggplot2::geom_tile plot of features by sample.
It is able to deal with multiple types affecting the same sample.
}
\examples{
v1 <- c("RCOR1", "NCOR1", "LCOR", "RCOR1", "RCOR1", "RCOR1", "RCOR1")
v2 <- c("sampleA", "sampleC", "sampleB", "sampleC", "sampleA", "sampleC", "sampleC")
v3 <- c("Deletion", "Deletion", "SNV", "Rearrangement", "SNV", "Rearrangement", "SNV")
feature.order <- c("RCOR1", "NCOR1", "LCOR")
sample.id.order <- c("sampleA", "sampleB", "sampleC")
in.dt <- data.table::data.table(feature = v1, sampleID = v2, type = v3)
fill.colors <- c("Deletion" = "Blue", "Rearrangement" = "Green", "SNV" = "Red")

plot_feature_sample_mat(in.dt)

# With black tile color
plot_feature_sample_mat(in.dt, tile.col = "black")

# Rotate x-axes labels by 90 degrees
plot_feature_sample_mat(in.dt, rotate.x.labels = 90)

# Specify order of features, samples, and colors
plot_feature_sample_mat(in.dt, feature.order, sample.id.order,
  fill.colors = fill.colors)

# Specify each cell can only have one "feature type"
plot_feature_sample_mat(in.dt, feature.order, sample.id.order, fill.colors = fill.colors,
  type.display.mode = "single")

# Specify the specific priority of the "feature type" for cells with
# multiple features
plot_feature_sample_mat(in.dt, feature.order, sample.id.order, fill.colors = fill.colors,
  type.display.mode = "single", type.order = c("Rearrangement", "SNV", "Deletion"))
}

